requirements.txt
flask==3.0.0
flask-cors==4.0.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
numpy==1.24.3
diffusers==0.25.0
transformers==4.36.0
accelerate==0.25.0
safetensors==0.4.1

